Move get_output to global fixtures
